+++ frontend/src/editor/components/sprites/recording-ignored-sprite.tsx
@@ -9,8 +9,8 @@ const RecordingIgnoredSprite = ({ x, y }: { x: number; y: number }) => {
         background: `url('${new URL("../../img/ignored_square.png", import.meta.url).href}') top left no-repeat`,
         width: STAGE_CELL_SIZE,
         height: STAGE_CELL_SIZE,
-        left: x * STAGE_CELL_SIZE,
-        bottom: y * STAGE_CELL_SIZE,
+        left: (x - 1) * STAGE_CELL_SIZE,
+        bottom: (y - 1) * STAGE_CELL_SIZE,
       }}
     />
   );